+++ src/main/java/com/example/cbumanage/repository/ProblemRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ProblemRepository extends JpaRepository<Problem, Integer>,
+public interface ProblemRepository extends JpaRepository<Problem, Long>,
         JpaSpecificationExecutor<Problem> {
 }
